streamlit_chat/app/requirements.txt
streamlit
langchain
streamlit_chat
python-dotenv
langchain-openai
openai
qdrant-client